fix(Abstract Factory): 추상 팩토리 메소드 패턴
- ktlint
