--- zeri/gof/src/main/kotlin/com/gof/pattern/abstract_factory/after/Wheel.kt
@@ -1,5 +1,0 @@
-package com.gof.pattern.abstract_factory.after
-
-interface Wheel {
-
-}
