--- zeri/gof/src/main/kotlin/com/gof/pattern/abstract_factory/after/WhiteAnchorPro.kt
@@ -1,4 +1,0 @@
-package com.gof.pattern.abstract_factory.after
-
-class WhiteAnchorPro : Anchor {
-}
