--- zeri/gof/src/main/kotlin/com/gof/pattern/abstract_factory/before/WhiteAnchor.kt
@@ -1,7 +1,0 @@
-package com.gof.pattern.abstract_factory.before
-
-import com.gof.pattern.abstract_factory.after.Anchor
-
-
-class WhiteAnchor : Anchor {
-}
