--- zeri/gof/src/main/kotlin/com/gof/pattern/abstract_factory/before/WhiteWheel.kt
@@ -1,7 +1,0 @@
-package com.gof.pattern.abstract_factory.before
-
-import com.gof.pattern.abstract_factory.after.Wheel
-
-
-class WhiteWheel : Wheel{
-}
