+++ zeri/gof/src/main/kotlin/com/gof/pattern/abstractfactory/after/Wheel.kt
@@ -1,0 +1,5 @@
+package com.gof.pattern.abstractfactory.after
+
+interface Wheel {
+
+}
