+++ zeri/gof/src/main/kotlin/com/gof/pattern/abstractfactory/after/WhiteAnchorPro.kt
@@ -1,0 +1,4 @@
+package com.gof.pattern.abstractfactory.after
+
+class WhiteAnchorPro : Anchor {
+}
