+++ zeri/gof/src/main/kotlin/com/gof/pattern/abstractfactory/after/WhiteWheelPro.kt
@@ -1,0 +1,4 @@
+package com.gof.pattern.abstractfactory.after
+
+class WhiteWheelPro : Wheel {
+}
