+++ zeri/gof/src/main/kotlin/com/gof/pattern/abstractfactory/before/WhiteAnchor.kt
@@ -1,0 +1,7 @@
+package com.gof.pattern.abstractfactory.before
+
+import com.gof.pattern.abstractfactory.after.Anchor
+
+
+class WhiteAnchor : Anchor {
+}
