+++ zeri/gof/src/main/kotlin/com/gof/pattern/abstractfactory/before/WhiteWheel.kt
@@ -1,0 +1,7 @@
+package com.gof.pattern.abstractfactory.before
+
+import com.gof.pattern.abstractfactory.after.Wheel
+
+
+class WhiteWheel : Wheel{
+}
